requirements.txt
openai
pydantic
python-dotenv
rapidfuzz
numpy